Updated restaurant components
